refactor(inspector): extract shared elicitation form primitives
Deduplicate form-rendering logic between ElicitationRequestDisplay (panel)
and InlineElicitationCard (chat) by extracting a shared module:

- schemaHelpers: enum/one-of/any-of choice parsing
- useElicitationForm: form state, schema-default init, required-field validation
- ElicitationFormFields: parameterized field renderer (boolean, number,
  select, enum, multi-select, textarea, input)

Both callers now consume these pieces while keeping their distinct chrome
(panel header/footer/toast vs. compact card + collapsed-after-response).
Test IDs preserved so e2e coverage is unaffected.
