+++ libraries/typescript/packages/inspector/src/client/components/elicitation/shared/index.ts
@@ -1,0 +1,8 @@
+export { ElicitationFormFields } from "./ElicitationFormFields";
+export { useElicitationForm } from "./useElicitationForm";
+export {
+  getMultiSelectChoices,
+  getSingleSelectChoices,
+  isEnumChoice,
+} from "./schemaHelpers";
+export type { EnumChoice } from "./schemaHelpers";
